move methods to utils
