+++ scripts/tasks_retrival/Genecorpus_tasks_creation.py
@@ -29,15 +29,16 @@ Epub ahead of print. (*co-corresponding authors)
 
 """
 
-import pickle
-from io import BytesIO
 from pathlib import Path
 
 import click
-import mygene
 import pandas as pd
-import requests
-from task_retrieval import report_task_single_col, verify_source_of_data
+from task_retrieval import (
+    get_symbols,
+    load_pickle_from_url,
+    report_task_single_col,
+    verify_source_of_data,
+)
 
 from gene_benchmark.tasks import dump_task_definitions
 
@@ -50,80 +51,6 @@ DATA_FILE_NAMES = {
     "bivalent vs non-methylated": "bivalent_promoters/bivalent_vs_no_methyl.pickle?download=true",
 }
 DATA_URL = "https://huggingface.co/datasets/ctheodoris/Genecorpus-30M/resolve/main/example_input_files/gene_classification/"
-
-
-def load_pickle_from_url(url):
-    """
-    Load a pickle file from a URL.
-
-    Parameters
-    ----------
-    url (str): The URL of the pickle file.
-
-    Returns
-    -------
-    object: The object loaded from the pickle file.
-
-    """
-    try:
-        response = requests.get(url)
-        response.raise_for_status()  # Check if the request was successful
-
-        # Create a BytesIO object from the response content
-        file_object = BytesIO(response.content)
-
-        # Load the pickle file from the BytesIO object
-        data = pickle.load(file_object)
-
-        return data
-    except requests.exceptions.RequestException as e:
-        print(f"Error downloading the file: {e}")
-    except pickle.UnpicklingError as e:
-        print(f"Error loading the pickle file: {e}")
-
-
-def get_symbols(gene_targetId_list):
-    """
-        given s list of gene id's (names Like ENSG00000006468) this method
-        uses the MyGenInfo package to retrieve the gene symbol (name like PLAC4).
-
-    Args:
-    ----
-        gene_targetId_list (list): list of gene id's (names Like ENSG00000006468)
-
-    Returns:
-    -------
-        list: List of corresponding symbols
-
-    """
-    mg = mygene.MyGeneInfo()
-    list_of_gene_metadata = mg.querymany(
-        gene_targetId_list, species="human", fields="symbol"
-    )
-    gene_metadata_df = get_id_to_symbol_df(list_of_gene_metadata)
-    symblist = [gene_metadata_df.loc[x, "symbol"] for x in gene_targetId_list]
-    return [v for v in symblist if not pd.isna(v)]
-
-
-def get_id_to_symbol_df(list_of_gene_metadata):
-    """
-        The method converts a list of gene metadata into a data frame,
-        each dictionary will contain the field symbol and the gene id as the query value.
-
-    Args:
-    ----
-        list_of_gene_metadata (list): list containing gene metadata.
-
-    Returns:
-    -------
-        pd.DataFrame: a data frame with the gene id as index with the symbol as value
-
-    """
-    gene_metadata_df = pd.DataFrame(list_of_gene_metadata)
-    # some target id have multiple symbols
-    gene_metadata_df = gene_metadata_df.drop_duplicates(subset="query")
-    gene_metadata_df.index = gene_metadata_df["query"]
-    return gene_metadata_df
 
 
 def create_symbol_dict(data_dict):
